web_one/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mysql-connector-python==8.3.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
zipp==3.17.0
